+++ src/Stack/ConfigCmd.hs
@@ -18,6 +18,8 @@ module Stack.ConfigCmd
   , cfgCmdEnvName
   , cfgCmdBuildFiles
   , cfgCmdBuildFilesName
+  , cfgCmdCompilerTools
+  , cfgCmdCompilerToolsName
   , cfgCmdName
   , yamlContainsInclude
   ) where
@@ -32,33 +34,27 @@ import qualified Data.Map.Merge.Strict as Map
 import qualified Data.Text as T
 import qualified Data.Yaml as Yaml
 import           Pantry ( loadSnapshot )
-import           Path ( (</>), parent )
+import           Path ( parent )
+import           Path.IO ( ensureDir )
 import qualified RIO.Map as Map
 import           RIO.NonEmpty ( nonEmpty )
 import qualified RIO.NonEmpty as NE
 import           RIO.Process ( envVarsL )
-import           Stack.Config
-                   ( makeConcreteSnapshot, getProjectConfig
-                   , getImplicitGlobalProjectDir
-                   )
-import           Stack.Constants ( stackDotYaml )
+import           Stack.Config ( makeConcreteSnapshot, withConfigExtra )
 import           Stack.Prelude
 import           Stack.Types.BuildConfig ( BuildConfig )
 import           Stack.Types.Config ( Config (..), HasConfig (..) )
+import           Stack.Types.ConfigExtra ( ConfigExtra (..) )
 import           Stack.Types.ConfigMonoid
-                   ( configMonoidInstallGHCName
-                   , configMonoidInstallMsysName
+                   ( configMonoidInstallGHCName, configMonoidInstallMsysName
                    , configMonoidRecommendStackUpgradeName
                    , configMonoidSystemGHCName
                    )
 import           Stack.Types.ConfigSetOpts
                    ( CommandScope (..), ConfigCmdSet (..) ,configCmdSetScope )
-import           Stack.Types.EnvConfig ( EnvConfig )
+import           Stack.Types.EnvConfig ( EnvConfig, bindirCompilerTools )
 import           Stack.Types.EnvSettings ( EnvSettings (..) )
 import           Stack.Types.GHCVariant ( HasGHCVariant )
-import           Stack.Types.GlobalOpts ( GlobalOpts (..) )
-import           Stack.Types.ProjectConfig ( ProjectConfig (..) )
-import           Stack.Types.Runner ( globalOptsL )
 import           Stack.Types.Snapshot ( AbstractSnapshot )
 import           System.Environment ( getEnvironment )
 
@@ -95,18 +91,14 @@ instance Exception ConfigCmdPrettyException
 cfgCmdSet ::
      (HasConfig env, HasGHCVariant env)
   => ConfigCmdSet -> RIO env ()
-cfgCmdSet cmd = do
-  conf <- view configL
-  configFilePath <-
-    case configCmdSetScope cmd of
-      CommandScopeProject -> do
-        mstackYamlOption <- view $ globalOptsL . to (.stackYaml)
-        mstackYaml <- getProjectConfig mstackYamlOption
-        case mstackYaml of
-          PCProject stackYaml -> pure stackYaml
-          PCGlobalProject -> getImplicitGlobalProjectDir <&> (</> stackDotYaml)
-          PCNoProject _extraDeps -> prettyThrowIO NoProjectConfigAvailable
-          -- maybe modify the ~/.stack/config.yaml file instead?
+-- We ignore any user message in the project-level configuration file:
+cfgCmdSet cmd = withConfigExtra False $ \configExtra -> do
+  let conf = configExtra.config
+  configFilePath <- case configCmdSetScope cmd of
+      CommandScopeProject -> case configExtra.configFile of
+        Left _ -> prettyThrowIO NoProjectConfigAvailable
+        -- Maybe modify the global configuration file (config.yaml) instead?
+        Right fp -> pure fp
       CommandScopeGlobal -> pure conf.userGlobalConfigFile
   rawConfig <- liftIO (readFileUtf8 (toFilePath configFilePath))
   config <- either throwM pure (Yaml.decodeEither' $ encodeUtf8 rawConfig)
@@ -348,6 +340,10 @@ cfgCmdEnvName = "env"
 cfgCmdBuildFilesName :: String
 cfgCmdBuildFilesName = "build-files"
 
+-- | The name of Stack's @config@ command's @compiler-tools-bin@ subcommand.
+cfgCmdCompilerToolsName :: String
+cfgCmdCompilerToolsName = "compiler-tools"
+
 data EnvVarAction = EVASet !Text | EVAUnset
   deriving Show
 
@@ -381,3 +377,7 @@ cfgCmdEnv es = do
 -- 'Stack.Config.withBuildConfig' that yields the desired actions.
 cfgCmdBuildFiles :: () -> RIO BuildConfig ()
 cfgCmdBuildFiles () = pure ()
+
+-- | This function takes no settings.
+cfgCmdCompilerTools :: () -> RIO EnvConfig ()
+cfgCmdCompilerTools () = bindirCompilerTools >>= ensureDir