+++ tests/integration/tests/2643-copy-compiler-tool/files/appB/Main.hs
@@ -1,0 +1,4 @@
+module Main (main) where
+
+main :: IO ()
+main = pure ()