+++ tests/integration/tests/4181-clean-wo-dl-ghc/Main.hs
@@ -3,7 +3,7 @@
 --
 -- See: https://github.com/commercialhaskell/stack/issues/4181
 
-import StackTest
+import           StackTest
 
 main :: IO ()
 main = do