+++ tests/integration/tests/617-unused-flag-cli/Main.hs
@@ -4,7 +4,7 @@
 --
 -- See: https://github.com/commercialhaskell/stack/issues/617
 
-import StackTest
+import           StackTest
 
 main :: IO ()
 main = do